requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.0.12
Django==4.0.5
idna==3.3
python-dotenv==0.20.0
requests==2.28.0
sqlparse==0.4.2
urllib3==1.26.9
